requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
pyside6==6.9.0
numpy==1.26.4
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
sounddevice==0.5.2
soundcard==0.4.4
Halo
pyaudio
resampy==0.4.3
librosa==0.10.1
setuptools
tqdm
huggingface_hub
ctranslate2==4.6.0
faster-whisper==1.1.1
scipy==1.15.3
sentencepiece==0.1.99
nltk==3.9.1
google-generativeai